LPS-79679 Source formatting, no logical changes
